adjust 'run wlutz' to match desired API
